refactor(admin): fix errors and refactor
