+++ src/pages/Admin/TabPanel.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import {
+  Typography, Box,
+} from '@material-ui/core';
+
+export default function TabPanel(props) {
+  const {
+    children, value, index, ...other
+  } = props;
+
+  return (
+    <Typography
+      component="div"
+      role="tabpanel"
+      hidden={value !== index}
+      id={`vertical-tabpanel-${index}`}
+      aria-labelledby={`vertical-tab-${index}`}
+      {...other}
+    >
+      {value === index && <Box p={3}>{children}</Box>}
+    </Typography>
+  );
+}
