+++ src/pages/Admin/tabs.js
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+export const tabs = [
+  'Verificación comunitaria',
+  'Verificación en departamento',
+  'Verificación en juzgado',
+  'Todas',
+];
